xen/arm: Don't set txdone_poll flag for ARM SMC mailbox
Don't set txdone_poll flag resulting in TXDONE_BY_POLL method.
It is not optimal to use this method along with the dummy
last_tx_done(), since the controller is completely synchronous.
What is more the TXDONE_BY_POLL method is prohibited because of
involving timer based polling.

This change leads to using TXDONE_BY_ACK method and as the result
the client (SCPI protocol) explicitly ticks the TX state machine.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>