arm: add platform_domain_create/destroy hook
Introduce platform specific create/destroy hook.
Normally this is not recommended, because the domain
is binded with platform which could not migrated to other
platform.

But we are supporting IoT case with device passthrough,
we need to let the VM to be platform specific.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Leonard Crestez <leonard.crestez@nxp.com>
(cherry picked from commit b5c04b2c9836dd9977917132c8c47df7bd65e763)