cpufreq: hack: Add extra parameters for userspace governor
There is a possibility to set a single frequency for all CPUs
via command line.

But, it is not appropriate for BL systems where CPU clusters
may operate with different frequencies. So, add a few parameters
for the userspace governor to understand what exactly frequency
to use.

For example, the following string:
cpufreq=xen:userspace,speed=1500000,count=4,speed1=1200000
means that speed '1500000' will be used for the first '4' CPUs [0...3]
and speed1 '1200000' will be used for all other CPUs.

Please note, this patch should be reworked.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>