Do not overwrite domain resource while reading from device tree
Domain ID, which is used to find the next free index in the
domain configuration array, cannot be set at system initialization time
as no domains can be running at this early stage. Thus, logic
to detect where to put the configuration, read from the Xen device
tree, must not rely on domain ID and simply put it into the next free
slot. Otherwise it leads to resources are always written at index 1,
thus overwriting the previous contents.

To fix that initialize used slots with DOMID_INVALID which is then
used as occupancy flag and initializes domain ID to invalid value.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Reviewed-by: Andrii Anisov <andrii_anisov@epam.com>