[ARM CODE] xen/arm: Add mapping for ITS translation register when IOMMU is enabled
If ITS is enabled for hardware domain and devices that are behind an
SMMU generate an MSI interrupt, SMMU fault will be observed as there
will be no mapping in p2m table for the ITS translation register.

MMIO mapping is required in the p2m table so that the device can
generate the MSI interrupt writing to the GITS_TRASLATOR register.

Mapping the ITS translation register to the hardware domain.

Change-Id: I2cb8b7b8173573d847f433fae5fc133844fa1aee
Signed-off-by: Rahul Singh <rahul.singh@arm.com>