iommu/ipmmu-vmsa: Remove all unused register definitions
This is a prereq work needed to add support for S4 series easily
in the future.

This is a non-verbatim port of corresponding Linux upsteam commit:
77cf983 iommu/ipmmu-vmsa: Remove all unused register definitions

Although Linux and Xen drivers have a lot in common, the main
differences are in translation stages (table formats), VMSAv8 modes,
supported SoC generations, etc, therefore that's why there is
a slight difference in registers/bits each driver considers unused.

No change in behavior.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>