iommu/ipmmu-vmsa: Add light version of Linux's ipmmu_features
This is a prereq work needed to add support for S4 series easily
in the future.

We don't need to port the whole struct as Xen driver doesn't
support old Arm32 based Gen2 SoCs, so there is no point in keeping
all differences between Gen2 and Gen3. All what we will need is
a minimal support to be able to operate with Gen3 and new S4.

Add Gen3 specific info with only two fields (number_of_contexts and
num_utlbs) for now, the subsequent patches will add remaining bits.

No change in behavior.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>