[ARM CODE] xen/arm: Enable the existing MSI-X and MSI handlers support for ARM
The existing MSI-X and MSI handlers support available for X86 is adapted
for Arm.

Enable existing MSIX handlers for accesses to the MSI-X message control
field on the PCI configuration space, and traps for accesses to the
memory region that contains the MSI-X table and PBA for ARM. This traps
detect attempts from the guest to configure MSI-X interrupts and
properly sets them up.

Enable existing handlers for the MSI control, address, data and mask
fields in order to detect accesses to them and setup the interrupts as
requested by the guest.

Signed-off-by: Rahul Singh <rahul.singh@arm.com>