pci: add rwlock to pcidevs_lock machinery
Currently pcidevs lock is a global recursive spinlock which is fine for
the existing use cases. It is used to both protect pdev instances
themselves from being removed while in use and to make sure the update
of the relevant pdev properties is synchronized.

Moving towards vPCI is used for guests this becomes problematic in terms
of lock contention. For example, during vpci_{read|write} the access to
pdev must be protected to prevent pdev disappearing under our feet.
This needs to be done with the help of pcidevs_{lock|unlock}.
On the other hand it is highly undesirable to lock all other pdev accesses
which only use pdevs in read mode, e.g. those which do not remove or
add pdevs.

For the above reasons introduce a read/write lock which will help
preventing locking contentions between pdev readers and writers:
- make the following use the new write lock:
  - pci_{add|remove}_device
  - pci_release_devices
  - device_assigned, {de|}assign_device
  - AMD/VT-d IOMMUs
  - setup_hwdom_pci_devices
  - vpci_add_handlers, vpci_remove_device, vpci_{add|remove}_register
- keep all the rest using the existing API (pcidevs_{lock|unlock},
  but extend the later to also acquire the rwlock in read mode.

Add new locking helpers: pcidevs_read_{un|}lock, pcidevs_write_{un|}lock
and means to check if the lock is held in read or write mode.

Note, that pcidevs_read_{un}lock doesn't acquire _pcidevs_lock recursive
lock because its users are not expected to modify pdev's contents.

This is in preparation for vPCI to be used for guests.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>

---
Since v1:
- user per CPU variable to track recursive readers
- more under write lock: pci_release_devices, device_assigned,
  assign_device, deassign_device, IOMMUs and vpci