vpci: use pcidevs locking to protect MMIO handlers
vPCI MMIO handlers are accessing pdevs without protecting this access
with pcidevs_{lock|unlock}. This is not a problem as of now as these
are only used by Dom0. But, towards vPCI is used also for guests, we need
to properly protect pdev and pdev->vpci from being removed while still
in use.

For that use pcidevs_read_{un}lock helpers. Note, that pcidevs_read_{un}lock
doesn't acquire _pcidevs_lock recursive lock because its users are not
expected to modify pdev's contents other than pdev->vpci which is protected
by pdev->vpci->lock (where appropriate).

This patch adds ASSERTs in the code to check that the rwlock is taken
and in appropriate mode. Some of such checks require changes to the
initialization of local variables which may be accessed before the
ASSERT checks the locking. For example see init_bars and mask_write.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>

---
Since v1:
- move pcidevs_read_{lock|unlock} into patch 1