xen/arm: Implement suspend/resume for domain watchdog
While a domain is suspended its watchdogs must be paused. Otherwise,
if the domain stays in the suspend state for a longer period of time
compared to the watchdog period, the domain would be shutdown on resume.
Proper solution to this problem is to stop (suspend) the watchdog timers
after the domain suspends and to restart (resume) the watchdog timers
before the domain resumes. The suspend/resume of watchdog timers is done
in Xen and is invisible to the guests.
On resume the watchdog timers are restarted with a timeout value equal
to the last set timeout before the suspend. This makes the watchdog
waiting period up to 2 times longer for the cycle that was active during
the suspend. But AFAIK there is no real benefig in keeping the watchdog
periods precise across suspend and this approach ensures the guests have
enough time to properly wake up.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>